split InitOnce to its own separate file
