+++ src/pc/stages/stage_loader.h
@@ -31,18 +31,4 @@ SpritePart* LoadSpriteParts(const char* filePath);
 // reset all the pre-allocated arrays, useful when loading a new stage
 void LoadReset();
 
-extern const char* OVL_EXPORT(GoldCollectTexts)[10];
-static void InitOnce() {
-    static bool is_initialized = 0;
-    if (is_initialized) {
-        return;
-    }
-    is_initialized = 1;
-
-    for (int i = 0; i < LEN(OVL_EXPORT(GoldCollectTexts)); i++) {
-        OVL_EXPORT(GoldCollectTexts)
-        [i] = AnsiToSotnMenuString(OVL_EXPORT(GoldCollectTexts)[i]);
-    }
-}
-
 #endif
