+++ src/pc/stages/stage_nz0.c
@@ -1,10 +1,9 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include <game.h>
-#include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #include "../st/nz0/nz0.h"
 #include "stage_loader.h"
+#include "stage_init.h"
 
 extern AbbreviatedOverlay OVL_EXPORT(Overlay);
 extern PfnEntityUpdate OVL_EXPORT(EntityUpdates)[];
